+++ tests/testcore/functional_test_base_suite.go
@@ -40,7 +40,7 @@ type FunctionalTestBaseSuite struct {
 }
 
 func (s *FunctionalTestBaseSuite) SetupSuite() {
-	s.FunctionalTestBase.SetupSuite("../testdata/es_cluster.yaml",
+	s.FunctionalTestBase.SetupSuite("testdata/es_cluster.yaml",
 		WithFxOptionsForService(primitives.FrontendService, fx.Populate(&s.frontendServiceName)),
 		WithFxOptionsForService(primitives.MatchingService, fx.Populate(&s.matchingServiceName)),
 		WithFxOptionsForService(primitives.HistoryService, fx.Populate(&s.historyServiceName)),
